refactor(browser): modularize browser controller flow
