+++ packages/browser/src/manifest/globs.ts
@@ -1,0 +1,87 @@
+import * as picomatch from 'picomatch';
+
+/**
+ * Convert a single glob pattern to RegExp.
+ */
+export const globToRegexp = (glob: string): RegExp => {
+  const regex = picomatch.makeRe(glob, {
+    fastpaths: false,
+    noglobstar: false,
+    bash: false,
+  });
+
+  if (!regex) {
+    throw new Error(`Invalid glob pattern: ${glob}`);
+  }
+
+  // picomatch generates regex starting with ^
+  // For patterns starting with ./, we need special handling
+  if (!glob.startsWith('./')) {
+    return regex;
+  }
+
+  // makeRe is sort of funny. If you pass it a directory starting with `./` it
+  // creates a matcher that expects files with no prefix (e.g. `src/file.js`)
+  // but if you pass it a directory that starts with `../` it expects files that
+  // start with `../`. Let's make it consistent.
+  // Globs starting `**` need special treatment due to the regex they produce
+  return new RegExp(
+    [
+      '^\\.',
+      glob.startsWith('./**') ? '' : '[\\/]',
+      regex.source.substring(1),
+    ].join(''),
+  );
+};
+
+/**
+ * Convert rstest include glob patterns to RegExp.
+ */
+export const globPatternsToRegExp = (patterns: string[]): RegExp => {
+  const regexParts = patterns.map((pattern) => {
+    const regex = globToRegexp(pattern);
+    // Remove ^ anchor and $ anchor to allow combining patterns
+    let source = regex.source;
+    if (source.startsWith('^')) {
+      source = source.substring(1);
+    }
+    if (source.endsWith('$')) {
+      source = source.substring(0, source.length - 1);
+    }
+    return source;
+  });
+
+  return new RegExp(`(?:${regexParts.join('|')})$`);
+};
+
+/**
+ * Convert exclude patterns to a RegExp for import.meta.webpackContext's exclude option
+ * This is used at compile time to filter out files during bundling
+ */
+export const excludePatternsToRegExp = (patterns: string[]): RegExp | null => {
+  const keywords: string[] = [];
+  for (const pattern of patterns) {
+    // Extract the core part between ** wildcards
+    const match = pattern.match(
+      /\*\*\/\.?\{?([^/*{}]+(?:,[^/*{}]+)*)\}?\/?\*?\*?/,
+    );
+    if (match) {
+      // Handle {a,b,c} patterns
+      const parts = match[1]!.split(',');
+      for (const part of parts) {
+        // Clean up the part (remove leading dots for hidden dirs)
+        const cleaned = part.replace(/^\./, '');
+        if (cleaned && !keywords.includes(cleaned)) {
+          keywords.push(cleaned);
+        }
+      }
+    }
+  }
+
+  if (keywords.length === 0) {
+    return null;
+  }
+
+  // Create regex that matches paths containing these directory names
+  return new RegExp(`[\\/](${keywords.join('|')})[\\/]`);
+};
