+++ packages/browser/src/manifest/index.ts
@@ -1,0 +1,6 @@
+export { collectProjectEntries } from './entries';
+export { generateManifestModule } from './generateManifest';
+export {
+  getBrowserProjects,
+  getRuntimeConfigFromProject,
+} from './projectConfig';
